refactor: split hooks tests into max-lines compliant files
